+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.1] - 2026-07-02
+
+### Added
+
+- Publish Scoop manifest for Windows installs (#182) ([#182](https://github.com/szhekpisov/diffyml/pull/182))
+
+### Fixed
+
+- Apply path filters across documents in multi-doc YAML (#190) ([#190](https://github.com/szhekpisov/diffyml/pull/190))
+- Exclude deeply nested paths in bulk added/removed subtrees (#189) (#191) ([#191](https://github.com/szhekpisov/diffyml/pull/191))
+
 ## [1.7.0] - 2026-06-21
 
 ### Added
@@ -290,6 +301,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add GOPATH/bin PATH hint to Go Install section (#3) ([#3](https://github.com/szhekpisov/diffyml/pull/3))
 - Cleanup (#6) ([#6](https://github.com/szhekpisov/diffyml/pull/6))
 
+[1.7.1]: https://github.com/szhekpisov/diffyml/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/szhekpisov/diffyml/compare/v1.6.2...v1.7.0
 [1.6.2]: https://github.com/szhekpisov/diffyml/compare/v1.6.1...v1.6.2
 [1.6.1]: https://github.com/szhekpisov/diffyml/compare/v1.6.0...v1.6.1